Added spec_analysis_gui, README. Made edits in code.
In the continuum fitting and nav calculation code:
Added reset functionality using 'r' key.
Added discrete scale options
Fixed NI triplet issue (it was unable to distinguish between triplets when they were too close in wavelength and rounding off wavelengths, thus overwriting transitions)
Added contamination flag for the Nav window. This is saved in the master pickle file and readable by the other GUI.
In the spectral analysis (final summary) GUI:
Added the contamination part so now it shaded under contaminated Nav
Displays redshift of absorber on plot.
Saves and plots the average velocity of selected transitions as a guide.
Displays summary of active sightline and absorber only.
We can also zoom in and out in the Nav window (because python plot windows allow it anyway).